feat: add result page and redirect end page
